fix: added updateTag for revalidation
