+++ frameworks/nextjs/package/update-pages.ts
@@ -1,0 +1,7 @@
+"use server"
+
+import { updateTag } from "next/cache";
+
+export async function updatePages(tags: string[]) {
+    await Promise.all(tags.map((tag: string) => updateTag(tag)));
+}
